make query handler middleware generic for user aggregation and event aggregation
